+++ app/src/main/java/org/simple/clinic/AppDatabase.kt
@@ -80,7 +80,7 @@ import org.simple.clinic.util.room.UuidRoomTypeConverter
       OverdueAppointment::class,
       PatientSearchResult::class
     ],
-    version = 85,
+    version = 86,
     exportSchema = true
 )
 @TypeConverters(