+++ app/src/main/java/org/simple/clinic/feature/Feature.kt
@@ -27,5 +27,6 @@ enum class Feature(
   OverdueCount(true, "overdue_count"),
   VillageTypeAhead(false, "village_type_ahead"),
   InstantSearchQrCode(true, "instant_search_qr_code"),
-  EthiopianCalendar(true, "ethiopian_calendar")
+  EthiopianCalendar(true, "ethiopian_calendar"),
+  InstantSearchByPatientIdentifier(false, "instant_search_by_patient_identifier")
 }