+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # CHANGELOG
 
 ## Next Release
+### Features
+- Add Instant Search using Patient Identifiers (Bangladesh National id, Ethiopian medical record no., etc)
+
 ### Internal
 - Move QR code scanner preview view inside `ScanSimpleIdScreen`
 - Remove dialog & sheet from navigation back stack when dismissing
@@ -8,6 +11,7 @@
 - [In Progress: 20 Jan 2021] Material Theme-ing Migration
 - [In Progress: 08 Apr 2021] Add automatic performance profiling for the Room database queries
 - Add `IndiaNationalHealthId` as an identifier in `IdentifierType`
+- Add sql query for instant search by numeric criteria
 
 ### Changes
 - Rename `Scan BP passport` button to `Scan QR code`